Add verify_and_rehash_method argument
If this option is set it will add a method that verifies if a password
matches the known hash. If it does and the hash needs a rehash it will
rehash the password and store the result to the database. In either case
it returns the result of the verification. This method takes a password
as its single argument.